+++ Service.cs
@@ -9,11 +9,15 @@ using Dalamud.Game.Gui;
 using Dalamud.Game.Gui.Dtr;
 using Dalamud.IoC;
 using Dalamud.Plugin;
+using HaselTweaks.Utils;
 
 namespace HaselTweaks;
 
 public class Service
 {
+    public static GameFunctions GameFunctions { get; } = new();
+    public static StringUtils StringUtils { get; } = new();
+
     [PluginService] public static DalamudPluginInterface PluginInterface { get; private set; } = null!;
     [PluginService] public static ChatGui Chat { get; private set; } = null!;
     [PluginService] public static ClientState ClientState { get; private set; } = null!;