+++ Utils/ImGuiUtils.cs
@@ -31,8 +31,8 @@ public static class ImGuiUtils
 
     public static void DrawPaddedSeparator()
     {
-        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().FramePadding.Y);
+        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().ItemSpacing.Y);
         ImGui.Separator();
-        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().FramePadding.Y - 1);
+        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().ItemSpacing.Y - 1);
     }
 }